requirements.txt
requests==2.26.0
python-arango==7.5.0
gradio==3.34.0